Refactor all metronome functions back into metronome.js. #2
