+++ client/components/band.js
@@ -7,7 +7,7 @@ export default class RenderBand extends React.Component {
   }
   render() {
     return (
-      <div>
+      <div className="beat-row">
         <div className="band beatOne"></div>
         <div className="band beatTwo"></div>
         <div className="band beatThree"></div>
