+++ client/styles.css
@@ -1,8 +1,24 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  background-color: #34495e;
+}
+
+.beat-row {
+  text-align: center;
+}
+
 .band {
   display: inline-block;
+  margin: 20px;
   width: 250px;
   height: 250px;
   border: 3px solid black;
+  border-radius: 4px;
+  background-color: #95a5a6;
 }
 
 span {
